refactor : inputs the member's additional information, was changed from the auth package to the member package.
