+++ src/main/kotlin/com/project/scrumbleserver/controller/member/MemberController.kt
@@ -1,10 +1,10 @@
-package com.project.scrumbleserver.controller.auth
+package com.project.scrumbleserver.controller.member
 
-import com.project.scrumbleserver.api.auth.API_POST_MEMBER_INFO
-import com.project.scrumbleserver.api.auth.ApiPostMemberInfoRequest
+import com.project.scrumbleserver.api.member.API_POST_MEMBER_INFO
+import com.project.scrumbleserver.api.member.ApiPostMemberInfoRequest
 import com.project.scrumbleserver.global.api.ApiResponse
 import com.project.scrumbleserver.security.RequestUserRowid
-import com.project.scrumbleserver.service.auth.AuthService
+import com.project.scrumbleserver.service.member.MemberService
 import io.swagger.v3.oas.annotations.Operation
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.PostMapping
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class AuthController(
-    private val authService: AuthService,
+class MemberController(
+    private val memberService: MemberService
 ) {
 
     @Operation(
@@ -25,7 +25,7 @@ class AuthController(
         @RequestBody @Valid request: ApiPostMemberInfoRequest,
         @RequestUserRowid userRowid: Long,
     ): ApiResponse<String> {
-        authService.enterAdditionalInfo(request, userRowid)
+        memberService.enterAdditionalInfo(request, userRowid)
         return ApiResponse.empty()
     }
 }
